requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.7.1
astroid==2.0.4
attrs==17.4.0
Automat==0.6.0
autopep8==1.4.2
certifi==2018.1.18
cffi==1.12.3
chardet==3.0.4
click==6.7
cloudpickle==0.6.1
colorama==0.3.7
constantly==15.1.0
cryptography==2.1.4
csirtg-domainsml==0.0.0a6
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
editdistance==0.5.3
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
idna==2.6
imutils==0.5.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
lazy-object-proxy==1.3.1
Markdown==3.1
mock==3.0.5
numpy==1.16.2
opencv-contrib-python==3.4.4.19
pandas==0.23.4
Pillow==6.0.0
protobuf==3.7.1
pycodestyle==2.4.0
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
scikit-learn==0.19.2
scipy==1.0.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typed-ast==1.3.5
vulture==1.0
Werkzeug==0.15.4
wrapt==1.11.1
